requirements.txt
python-dotenv
langchain
langchain-community
tavily-python 